requirements.txt
flask==2.0.3
lxml==4.8.0
requests==2.27.1
graphql-server==3.0.0b5
graphql-core==3.2.0
selenium==4.1.3
immutable-config==1.0
webdriver-manager==3.5.3
pygments-graphql-lexer==0.1.0